--- a/src/main/java/seedu/duke/controllers/ModuleMethodsController.java
+++ b/src/main/java/seedu/duke/controllers/ModuleMethodsController.java
@@ -14,7 +14,6 @@ import static seedu.duke.controllers.ModuleServiceController.chooseToAddToSchedu
 import static seedu.duke.models.logic.Api.doesModuleExist;
 import static seedu.duke.models.logic.Api.getModulePrereqBasedOnCourse;
 import static seedu.duke.models.logic.MajorRequirements.printRequiredModules;
-import static seedu.duke.models.logic.ScheduleGenerator.generateRecommendedSchedule;
 import static seedu.duke.views.CommandLineView.displayMessage;
 import static seedu.duke.views.CommandLineView.displaySuccessfulAddMessage;
 import static seedu.duke.views.CommandLineView.showPrereqCEG;
@@ -93,7 +92,8 @@ public class ModuleMethodsController {
 
     public static void recommendScheduleToStudent(Student student) {
         CommandLineView.displayMessage("Hold on a sec! Generating your recommended schedule <3....");
-        ArrayList<String> recommendedSchedule = generateRecommendedSchedule(student.getMajor());
+        //to refactor
+        ArrayList<String> recommendedSchedule = student.getSchedule().generateRecommendedSchedule(student.getMajor());
         chooseToAddToSchedule(student, recommendedSchedule);
     }