--- a/src/main/java/seedu/duke/models/logic/Api.java
+++ b/src/main/java/seedu/duke/models/logic/Api.java
@@ -30,6 +30,7 @@ import seedu.duke.utils.errors.UserError;
 import seedu.duke.views.ModuleInfoView;
 
 
+
 public class Api {
 
     /**
@@ -82,8 +83,6 @@ public class Api {
     private static ArrayList<String> getExemptedPrerequisite(String moduleCode) {
         HashMap<String, ArrayList<String>> map = new HashMap<>();
         ArrayList<String> list1 = new ArrayList<>();
-        list1.add("MA1511");
-        list1.add("MA1512");
         map.put("CS1231", list1);
 
         ArrayList<String> list2 = new ArrayList<>();
@@ -240,6 +239,29 @@ public class Api {
     }
 
     /**
+     * Retrieves the requirements the module fulfills
+     *
+     * @author rohitcube
+     * @param moduleCode The module code to retrieve workload information for.
+     * @return A JSONArray containing workload details.
+     *
+     */
+    public static ArrayList<String> getModuleFulfilledRequirements(String moduleCode) {
+        try {
+            JSONObject moduleInfo = getFullModuleInfo(moduleCode);
+            ArrayList<String> fulfilledArray = new ArrayList<>();
+            ArrayList<String> response = (ArrayList<String>) moduleInfo.get("fulfillRequirements");
+            if(response != null){
+                fulfilledArray = response;
+            }
+
+            return fulfilledArray;
+        } catch (ClassCastException | NullPointerException e) {
+            return new ArrayList<String>();
+        }
+    }
+
+    /**
      * Recursively flattens and processes a list of module prerequisites.
      *
      * @author ryanlohyr
@@ -257,16 +279,23 @@ public class Api {
             ArrayList<String> courseRequirements,
             String currRequisite) {
         try {
+            int lengthOfModulePreReqArray = modulePrereqArray.size();
+            int counter = 0;
             for (Object module : modulePrereqArray) {
                 if (module instanceof String) {
-                    String formattedModule = ((String) module).replace(":D", "");
+                    String formattedModule = ((String) module).split(":")[0];
                     formattedModule = formattedModule.replace("%", "");
-
                     if (courseRequirements.contains(formattedModule)) {
                         prerequisites.add(formattedModule);
                         if (currRequisite.equals("or")) {
                             return;
                         }
+                    }
+                    //if this is the last item and the module also part of the courseRequirements, we add it anw
+                    if (currRequisite.equals("or") && counter == (lengthOfModulePreReqArray - 1)
+                            && !courseRequirements.contains((formattedModule))){
+                        prerequisites.add(formattedModule);
+                        return;
                     }
                 } else {
                     //item is an object
@@ -278,18 +307,22 @@ public class Api {
                         ArrayList<ArrayList<Objects>> initial = (ArrayList<ArrayList<Objects>>) moduleJSON.get("nOf");
                         ArrayList<Objects> formattedInitial = initial.get(1);
                         flattenPrereq(major, prerequisites, formattedInitial, courseRequirements, key);
-                        return;
+                    }else{
+                        String key = (String) moduleJSON.keySet().toArray()[0];
+
+                        ArrayList<Objects> initial = (ArrayList<Objects>) moduleJSON.get(key);
+
+                        flattenPrereq(major, prerequisites, initial, courseRequirements, key);
                     }
-                    String key = (String) moduleJSON.keySet().toArray()[0];
 
-                    ArrayList<Objects> initial = (ArrayList<Objects>) moduleJSON.get(key);
-
-                    flattenPrereq(major, prerequisites, initial, courseRequirements, key);
 
                 }
+                counter += 1;
             }
         } catch (ClassCastException e) {
-            System.out.println("Error getting pre requisite for module");
+            System.out.println("Sorry but we could not get the prerequisite " +
+                    "for this module as NUSMods API provided it " +
+                    "in a format we did not expect :<");
         }
     }
 
@@ -303,7 +336,6 @@ public class Api {
      *
      */
     public static ArrayList<String> getModulePrereqBasedOnCourse(String moduleCode, String major) {
-        // Only accepts CEG requirements now
         try {
             Major.valueOf(major.toUpperCase());
         } catch (IllegalArgumentException e) {