--- a/src/main/java/seedu/duke/models/schema/Module.java
+++ b/src/main/java/seedu/duke/models/schema/Module.java
@@ -23,22 +23,21 @@ public class Module {
         if (moduleCode.isEmpty()) {
             throw new NullPointerException();
         }
-        try {
-            JSONObject response = getFullModuleInfo(moduleCode);
+        this.moduleCode = moduleCode;
+    }
 
-            assert response != null: "Response from NUSMods API is null";
-            assert !response.isEmpty(): "Response Object is empty";
-            this.moduleCode = moduleCode;
-            this.isCompleted = false;
-            this.moduleDescription = (String) response.get("description");
-            this.moduleName = (String) response.get("title");
-            try{
-                this.moduleCredits = (Integer) response.get("moduleCredit");
-            }catch (ClassCastException e){
-                this.moduleCredits = 4;
-            }
-        } catch (RuntimeException e) {
-            throw e;
+    public void loadNUSModsAPI(){
+        JSONObject response = getFullModuleInfo(moduleCode);
+        assert response != null: "Response from NUSMods API is null";
+        assert !response.isEmpty(): "Response Object is empty";
+
+        this.isCompleted = false;
+        this.moduleDescription = (String) response.get("description");
+        this.moduleName = (String) response.get("title");
+        try{
+            this.moduleCredits = (Integer) response.get("moduleCredit");
+        }catch (ClassCastException e){
+            this.moduleCredits = 4;
         }
     }