--- a/src/main/java/seedu/duke/models/schema/ModuleList.java
+++ b/src/main/java/seedu/duke/models/schema/ModuleList.java
@@ -26,7 +26,6 @@ public class ModuleList {
 
     public ModuleList(String modules) {
         this();
-
         if (modules == null || modules.isEmpty()) {
             return;
         }