--- a/src/main/java/seedu/duke/models/schema/Schedule.java
+++ b/src/main/java/seedu/duke/models/schema/Schedule.java
@@ -1,13 +1,19 @@
 package seedu.duke.models.schema;
 
 import seedu.duke.exceptions.FailPrereqException;
+import seedu.duke.exceptions.MissingModuleException;
 
 import java.io.InvalidObjectException;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
-
+import java.util.Queue;
+import java.util.LinkedList;
+import static seedu.duke.models.logic.Api.getModulePrereqBasedOnCourse;
+import static seedu.duke.models.logic.Api.getModuleFulfilledRequirements;
 import static seedu.duke.models.logic.Api.doesModuleExist;
 import static seedu.duke.models.logic.Api.satisfiesAllPrereq;
+import static seedu.duke.models.logic.DataRepository.getRequirements;
 import static seedu.duke.views.SemesterPlannerView.printSemesterPlanner;
 
 /**
@@ -20,6 +26,9 @@ public class Schedule {
     protected int[] modulesPerSem;
     private ModuleList modulesPlanned;
 
+    private HashMap<String, ArrayList<String>> prereqMap;
+
+
     /**
      * Constructs a new `Schedule` with the provided modules and distribution across semesters.
      *
@@ -29,6 +38,7 @@ public class Schedule {
     public Schedule(String modules, int[] modulesPerSem) {
         modulesPlanned = new ModuleList(modules);
         this.modulesPerSem = modulesPerSem;
+
     }
 
     /**
@@ -69,12 +79,14 @@ public class Schedule {
             }
 
             //Sub list as we only want modules before the current target semester
-            List<String> completedModulesArray = scheduleToAdd.subList(0, (indexToAdd));
-            ModuleList completedModules = new ModuleList(String.join(" ", completedModulesArray));
-
-            if (!satisfiesAllPrereq(module, completedModules)){
-                currentSem += 1;
-                currentIndexOfMod = 0;
+            List<String> currentSemestersModules = scheduleToAdd.subList(indexToAdd, indexToAdd + currentIndexOfMod);
+            ArrayList<String> currModulesPrereq = prereqMap.get(module);
+            //now we check if the modules prereq is contained on current line
+            for(String currModule:currentSemestersModules){
+                if(currModulesPrereq.contains(currModule)){
+                    currentSem += 1;
+                    currentIndexOfMod = 0;
+                }
             }
 
             try {
@@ -149,17 +161,13 @@ public class Schedule {
      * @throws FailPrereqException If the module to be deleted is a prerequisite for other modules in the schedule.
      * @throws IllegalArgumentException If the provided module code is not valid, the module is not in the schedule
      */
-    public void deleteModule(String module) throws FailPrereqException, IllegalArgumentException {
-        //int targetIndex = getMainModuleList().indexOf(module);
-        int targetIndex = modulesPlanned.getIndex(module);
+    public void deleteModule(String module) throws FailPrereqException, MissingModuleException {
 
         if (!doesModuleExist(module)) {
-            throw new IllegalArgumentException("Please select a valid module");
+            throw new MissingModuleException("Module does not exist in schedule");
         }
 
-        if (targetIndex == -1) {
-            throw new IllegalArgumentException("Module is not in schedule");
-        }
+        int targetIndex = modulesPlanned.getIndex(module);
 
         int targetSem = 1;
         int moduleCount = modulesPerSem[0];
@@ -169,35 +177,60 @@ public class Schedule {
             targetSem += 1;
         }
 
-        int nextSemStartingIndex = moduleCount;
-
-        int lastModuleIndex = modulesPlanned.getMainModuleList().size() - 1;
-
-        List<String> completedModulesArray = modulesPlanned.getModuleCodes().subList(0, nextSemStartingIndex);
-        ModuleList completedModules = new ModuleList(String.join(" ", completedModulesArray));
-        completedModules.deleteModulebyCode(module);
+        ArrayList<String> requirementsFulfilledFromModule = getModuleFulfilledRequirements(module);
 
         List<String> modulesAheadArray;
+        int lastModuleIndex = modulesPlanned.getMainModuleList().size() - 1;
+        int nextSemStartingIndex = moduleCount;
+
         try {
             modulesAheadArray = modulesPlanned.getModuleCodes().subList(nextSemStartingIndex, lastModuleIndex + 1);
         } catch (IndexOutOfBoundsException | IllegalArgumentException e) {
             modulesAheadArray = new ArrayList<>();
         }
 
-        try {
-            for (String moduleAhead : modulesAheadArray){
-                if (!satisfiesAllPrereq(moduleAhead, completedModules)) {
-                    throw new FailPrereqException("Unable to delete module. This module is a prerequisite for "
-                            + moduleAhead);
-                }
+        for(String fulfilledModule: requirementsFulfilledFromModule ){
+            //over here we check if the semesters in front of us contain a module in fulfilled module
+            if(modulesAheadArray.contains(fulfilledModule)){
+                throw new FailPrereqException("Unable to delete module. This module is a prerequisite for "
+                        + fulfilledModule);
             }
-        } catch (IllegalArgumentException e) {
-            // This catch should never occur as it should not be possible to add an invalid module
-            assert false;
-            throw new IllegalArgumentException("Invalid Module in Schedule");
         }
 
+        modulesPlanned.deleteModulebyCode(module);
+
         modulesPerSem[targetSem - 1] -= 1;
+
+        //        completedModules.deleteModulebyCode(module);
+        //        int nextSemStartingIndex = moduleCount;
+        //
+        //        int lastModuleIndex = modulesPlanned.getMainModuleList().size() - 1;
+        //        List<String> completedModulesArray = modulesPlanned.getModuleCodes().subList(0, nextSemStartingIndex);
+        //        ModuleList completedModules = new ModuleList(String.join(" ", completedModulesArray));
+        //        completedModules.deleteModulebyCode(module);
+        //
+        //        List<String> modulesAheadArray;
+        //        try {
+        //       modulesAheadArray = modulesPlanned.getModuleCodes().subList(nextSemStartingIndex, lastModuleIndex + 1);
+        //        } catch (IndexOutOfBoundsException | IllegalArgumentException e) {
+        //            modulesAheadArray = new ArrayList<>();
+        //        }
+        //
+        //        try {
+        //            for (String moduleAhead : modulesAheadArray){
+        //                if (!satisfiesAllPrereq(moduleAhead, completedModules)) {
+        //                   throw new FailPrereqException("Unable to delete module. This module is a prerequisite for "
+        //                            + moduleAhead);
+        //                }
+        //            }
+        //        } catch (IllegalArgumentException e) {
+        //            // This catch should never occur as it should not be possible to add an invalid module
+        //            assert false;
+        //            throw new IllegalArgumentException("Invalid Module in Schedule");
+        //        }
+
+        //        modulesPerSem[targetSem - 1] -= 1;
+        //
     }
 
     public Module getModule(String moduleCode) throws InvalidObjectException {
@@ -244,5 +277,65 @@ public class Schedule {
 
     public void printMainModuleList() {
         printSemesterPlanner(modulesPerSem, modulesPlanned.getMainModuleList());
+    }
+
+    /**
+     * Generates a recommended schedule for a given course based on its requirements and prerequisites.
+     *
+     * @author ryanlohyr
+     * @param course The course for which to generate a recommended schedule.
+     * @return An ArrayList of strings representing the recommended schedule in order of completion.
+     */
+    public ArrayList<String> generateRecommendedSchedule(String course){
+        ArrayList<String> requirements = getRequirements(course);
+        HashMap<String, Integer> degreeMap = new HashMap<>();
+        Queue<String> q = new LinkedList<>();
+        ArrayList<String> schedule = new ArrayList<>();
+        HashMap<String, ArrayList<String>> adjacencyList = new HashMap<>();
+        this.prereqMap = new HashMap<>();
+        //initialisation
+        for(String requirement: requirements) {
+            adjacencyList.put(requirement, new ArrayList<>());
+            degreeMap.put(requirement, 0);
+        }
+
+        for (String requirement : requirements) {
+            ArrayList<String> prereqArray = getModulePrereqBasedOnCourse(requirement, course);
+            if (prereqArray == null) {
+                prereqArray = new ArrayList<>();
+            }
+
+            prereqMap.put(requirement, prereqArray);
+
+            //we need to create an adjacency list to add all the connections
+            // from pre req -> item
+            for (String s : prereqArray) {
+                adjacencyList.get(s).add(requirement);
+                Integer value = degreeMap.get(requirement) + 1;
+                degreeMap.put(requirement, value);
+            }
+        }
+
+        for (String key : degreeMap.keySet()) {
+            Integer value = degreeMap.get(key);
+            if(value == 0){
+                q.offer(key);
+            }
+        }
+
+        while(!q.isEmpty()){
+            String curr = q.poll();
+            schedule.add(curr);
+            ArrayList<String> currReq = adjacencyList.get(curr);
+            for (String s : currReq) {
+                int num = degreeMap.get(s) - 1;
+                degreeMap.put(s, num);
+                if (num == 0) {
+                    q.offer(s);
+                }
+            }
+        }
+
+        return schedule;
     }
 }