… src/main/java/seedu/duke/models/logic/ScheduleGenerator.java