### Optimised add recommended schedule
function optimised from 150s to 1s :D ,

**_Original Method_**
Original function checks whether all pre requisites of the modules have been 

_**Optimised Method**_

1. We create an additional adjacency list during the generation of the recommended schedule, decreases the number of API calls, 
2. We use that adjacency list to check only the row in which we are adding the modules. (Not necessary to check the entire array as if the pre requisite is not in the same row, it is assumed it is valid to place it in that row as we have performed topological sort already)

### Optimised Delete function 
function optimised from 20s to 3s.

_**The original method**_ 
implemented checked every module ahead of its pre requisites to determine whether it was able to be deleted. 

_**Optimised Method**_
Optimised algorithm to check the modules it unlocks instead, and checks whether the modules ahead contains the modules unlocked. Significantly decreasing deletion time

The above 2 optimisations lead Github actions to be reduced from 18minutes -> 3 minutes.

### PED Bug Fixes
This pull request fixes #128 #127


### To note 
To reduce runtime of add modules to schedule, I removed the api in the constructor of `src/main/java/seedu/duke/models/schema/Module.java` and placed it in `loadNUSModsAPi` as such 

```
public void loadNUSModsAPI(){
        JSONObject response = getFullModuleInfo(moduleCode);
        assert response != null: "Response from NUSMods API is null";
        assert !response.isEmpty(): "Response Object is empty"; `

        this.isCompleted = false;
        this.moduleDescription = (String) response.get("description");
        this.moduleName = (String) response.get("title");
        try{
            this.moduleCredits = (Integer) response.get("moduleCredit");
        }catch (ClassCastException e){
            this.moduleCredits = 4; 
```

So everytime you want to access certain properties such as module credits or location for your timetable, you have to call it as such first 


```
Module module = Module("CS1010) 
module.loadNUSModsAPI() 
module.getModuleCredits();
```


However this implementation can be altered in the future, can just discuss